+++ crates/umasim/src/utils.rs
@@ -457,7 +457,6 @@ pub(crate) fn fallback_override_game_config() -> OverrideGameConfig {
         mcts: OverrideMctsConfig::default(),
         ramen_region_strategy: None,
         ramen_region_fixed: None,
-        ramen_pt_sacrifice_score: None
     }
 }